requirements.txt
Pillow>=10.0.0
tkinterdnd2>=0.3.0
onnxruntime>=1.16.0
numpy>=1.24.0
rembg>=2.0.0
pytest>=8.0.0
